+++ src/ExternalConnections/Payment/Payment.java
@@ -1,11 +1,9 @@
 package ExternalConnections.Payment;
 
-import ExternalConnections.ExternalConnections;
-
 public interface Payment {
 
     // 0 is success, 1 is connection invalid, 2 is insufficient funds
-    int payment (int total);
+    int payment (float total);
 
     // use secure key to connect
     boolean connect (int key);
