+++ src/ExternalConnections/Payment/Visa.java
@@ -15,7 +15,7 @@ public class Visa implements Payment {
     }
 
     @Override
-    public int payment(int total) {
+    public int payment(float total) {
         synchronized (lock) {
 
             return 0;
